Reverted.
svn path=/trunk/vorbis-tools/; revision=14482

+++ ogg123/ogg123.h
@@ -25,8 +25,8 @@
 typedef struct ogg123_options_t {
   long int verbosity;         /* Verbose output if > 1, quiet if 0 */
 
-  long int shuffle;           /* Should we shuffle playing? */
-  long int repeat;            /* Repeat playlist indefinitely? */
+  int shuffle;                /* Should we shuffle playing? */
+  int repeat;                 /* Repeat playlist indefinitely? */
   ogg_int64_t delay;          /* delay (in millisecs) for skip to next song */
   int nth;                    /* Play every nth chunk */
   int ntimes;                 /* Play every chunk n times */
